Refactors most private methods as PORO services
Commit: Ref

TaskNumber: -
